backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pytest==8.4.1
httpx==0.28.1
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
